requirements.txt
streamlit==1.31.0
sqlalchemy==2.0.25
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.0
